+++ spinta/cli/helpers/admin/components.py
@@ -16,3 +16,4 @@ class Script(enum.Enum):
     DEDUPLICATE = "deduplicate"
     CHANGELOG = "changelog"
     ENUM_LIST = "enum_list"
+    CITUS_DISTRIBUTION = "citus_distribution"