+++ CHANGES.rst
@@ -17,6 +17,13 @@ Improvements:
   - Invalid `prepare` expressions raise a clear `ModelNotFound` `PropertyNotFound`
 error pointing at the offending scope.
 
+New Features:
+
+- Added `spinta admin citus_distribution` script, that will apply configured distribution strategy (`#1691`_).
+- Added ability to configure default and model specific distribution strategy (`#1691`_).
+
+.. _#1691: https://github.com/atviriduomenys/spinta/issues/1691
+
 Bug fixes:
 
 - Fixed a bug where nested backrefs where causing an error (`#1608`_).